+++ libs/message.ts
@@ -8,12 +8,42 @@ dayjs.extend(timezone);
 dayjs.tz.setDefault("Asia/Tokyo");
 
 /**
+ * 返信メッセージを作成
+ * @param text テキスト
+ * @return Flexメッセージ
+ */
+export function createReplyMessage(text: string): FlexMessage {
+  const lines = text.split("\n");
+
+  // 行数が足りない
+  if (lines.length <= 1) {
+    return createHelpMessage();
+  }
+
+  // 走行距離・給油量
+  const mileage = parseFloat(lines[0]);
+  const refillQuantity = parseFloat(lines[1]);
+
+  // 数値ではない
+  if (isNaN(mileage) || isNaN(refillQuantity)) {
+    return createErrorMessage("数値エラー", "半角数字を入力してください");
+  }
+
+  // 数値が1以下
+  if (mileage <= 0 || refillQuantity <= 0) {
+    return createErrorMessage("計算エラー", "1以上の数値を入力してください");
+  }
+
+  return calcFuel(mileage, refillQuantity);
+}
+
+/**
  * 燃費の計算結果メッセージを作成
  * @param mileage 走行距離
  * @param refillQuantity 給油量
  * @returns Flexメッセージ
  */
-export function createFuelMessage(
+export function calcFuel(
   mileage: number,
   refillQuantity: number,
 ): FlexMessage {
